delete MetaServiceLegacy
