+++ src/Services/AssistantService.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Bavix\Wallet\Services;
 
+use Bavix\Wallet\Interfaces\CartInterface;
+use Bavix\Wallet\Interfaces\ProductInterface;
 use Bavix\Wallet\Internal\Dto\TransactionDtoInterface;
 use Bavix\Wallet\Internal\Dto\TransferDtoInterface;
 use Bavix\Wallet\Internal\Service\MathServiceInterface;
@@ -43,5 +45,23 @@ final class AssistantService implements AssistantServiceInterface
         }
 
         return array_filter($amounts, fn (string $amount): bool => $this->mathService->compare($amount, 0) !== 0);
+    }
+
+    public function getMeta(CartInterface $cart, ProductInterface $product): ?array
+    {
+        $metaCart = $cart->getBasketDto()
+            ->meta()
+        ;
+        $metaProduct = $product->getMetaProduct();
+
+        if ($metaProduct !== null) {
+            return array_merge($metaCart, $metaProduct);
+        }
+
+        if (count($metaCart) > 0) {
+            return $metaCart;
+        }
+
+        return null;
     }
 }
