+++ src/Services/AssistantServiceInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Bavix\Wallet\Services;
 
+use Bavix\Wallet\Interfaces\CartInterface;
+use Bavix\Wallet\Interfaces\ProductInterface;
 use Bavix\Wallet\Internal\Dto\TransactionDtoInterface;
 use Bavix\Wallet\Internal\Dto\TransferDtoInterface;
 
@@ -22,4 +24,6 @@ interface AssistantServiceInterface
      * @return array<int, string>
      */
     public function getSums(array $transactions): array;
+
+    public function getMeta(CartInterface $cart, ProductInterface $product): ?array;
 }
